backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.12.1
aiosqlite==0.19.0
numpy==1.26.2
scikit-learn==1.3.2
geoalchemy2==0.14.2

# AI推理依赖
ultralytics==8.0.200
torch==2.1.1
torchvision==0.16.1
opencv-python-headless==4.8.1.78

# 测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
